Split renderer/project-preview.tsx — extract file-tree and timeline parts
